# be/docs/openapi/event_analyzer.json
{"openapi":"3.0.0","paths":{"/analysis/event-forward":{"post":{"operationId":"AnalyzerController_receive","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","description":"AnalyzerPayload","additionalProperties":true}}}},"responses":{"202":{"description":"Accepted"},"204":{"description":"No Content"}},"summary":"이벤트를 수신하고 분석 파이프라인을 시작","tags":["analysis"]}},"/analysis":{"get":{"operationId":"AnalyzerController_getAllAnalysis","parameters":[{"name":"summary","required":false,"in":"query","schema":{"example":"timeout"}},{"name":"severity","required":false,"in":"query","schema":{"example":"high"}},{"name":"func","required":false,"in":"query","schema":{"example":"HW"}},{"name":"eventIds","required":false,"in":"query","description":"쉼표로 구분된 이벤트 ID 목록","schema":{"example":"101,102"}},{"name":"count","required":false,"in":"query","schema":{"example":50}},{"name":"startIndex","required":false,"in":"query","schema":{"example":0}},{"name":"end","required":false,"in":"query","schema":{"example":"2026-05-30"}},{"name":"start","required":false,"in":"query","schema":{"example":"2026-05-01"}}],"responses":{"200":{"description":"분석 결과 목록과 페이지 정보 반환"}},"summary":"분석 결과 목록을 조회","tags":["analysis"]}},"/analysis/{eventId}":{"get":{"operationId":"AnalyzerController_getAnalysis","parameters":[{"name":"eventId","required":true,"in":"path","schema":{"example":"101","type":"string"}}],"responses":{"200":{"description":"분석 단건 반환"}},"summary":"eventId 기준 분석 결과를 조회","tags":["analysis"]}},"/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":"헬스 체크 응답 반환"}},"summary":"서비스 헬스 상태를 조회","tags":["health"]}},"/query/logs":{"get":{"operationId":"QueryController_getLogs","parameters":[{"name":"count","required":false,"in":"query","schema":{"example":10}},{"name":"startIndex","required":false,"in":"query","schema":{"example":0}},{"name":"summary","required":false,"in":"query","schema":{"example":"timeout"}},{"name":"func","required":false,"in":"query","schema":{"example":"HW"}},{"name":"severity","required":false,"in":"query","schema":{"example":"high"}},{"name":"status","required":false,"in":"query","schema":{"example":"prepared"}},{"name":"end","required":false,"in":"query","schema":{"example":"2026-05-30"}},{"name":"start","required":false,"in":"query","schema":{"example":"2026-05-01"}}],"responses":{"200":{"description":"통합 로그 목록 반환 { code, data: items[], pageInfo: { totalCount, count, index, hasNext } }"}},"summary":"이벤트/분석 통합 로그 목록 조회","tags":["query"]}}},"info":{"title":"event_analyzer API","description":"Auto-generated API docs for event_analyzer","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{}}}